+++ sprockets/mixins/mediatype/handlers.py
@@ -7,14 +7,20 @@ Basic content handlers.
   to text before calling functions that encode & decode text
 
 """
+from __future__ import annotations
+
+import typing
+
 from tornado import escape
+
+from sprockets.mixins.mediatype import type_info
 
 
 class BinaryContentHandler:
     """
     Pack and unpack binary types.
 
-    :param str content_type: registered content type
+    :param content_type: registered content type
     :param pack: function that transforms an object instance
         into :class:`bytes`
     :param unpack: function that transforms :class:`bytes`
@@ -24,17 +30,21 @@ class BinaryContentHandler:
     and unpacking functions.
 
     """
-    def __init__(self, content_type, pack, unpack):
+    def __init__(self, content_type: str, pack: type_info.PackBFunction,
+                 unpack: type_info.UnpackBFunction) -> None:
         self._pack = pack
         self._unpack = unpack
         self.content_type = content_type
 
-    def to_bytes(self, inst_data, encoding=None):
+    def to_bytes(
+            self,
+            inst_data: type_info.Serializable,
+            encoding: typing.Optional[str] = None) -> typing.Tuple[str, bytes]:
         """
         Transform an object into :class:`bytes`.
 
-        :param object inst_data: object to encode
-        :param str encoding: ignored
+        :param inst_data: object to encode
+        :param encoding: ignored
         :returns: :class:`tuple` of the selected content
             type and the :class:`bytes` representation of
             `inst_data`
@@ -42,15 +52,15 @@ class BinaryContentHandler:
         """
         return self.content_type, self._pack(inst_data)
 
-    def from_bytes(self, data_bytes, encoding=None):
+    def from_bytes(
+            self,
+            data_bytes: bytes,
+            encoding: typing.Optional[str] = None) -> type_info.Deserialized:
         """
         Get an object from :class:`bytes`
 
-        :param bytes data_bytes: stream of bytes to decode
-        :param str encoding: ignored
-        :param dict content_parameters: optional :class:`dict` of
-            content type parameters from the :mailheader:`Content-Type`
-            header
+        :param data_bytes: stream of bytes to decode
+        :param encoding: ignored
         :returns: decoded :class:`object` instance
 
         """
@@ -61,12 +71,12 @@ class TextContentHandler:
     """
     Transcodes between textual and object representations.
 
-    :param str content_type: registered content type
+    :param content_type: registered content type
     :param dumps: function that transforms an object instance
         into a :class:`str`
     :param loads: function that transforms a :class:`str`
         into an object instance
-    :param str default_encoding: encoding to apply when
+    :param default_encoding: encoding to apply when
         transcoding from the underlying body :class:`byte`
         instance
 
@@ -76,18 +86,23 @@ class TextContentHandler:
     that tornado expects.
 
     """
-    def __init__(self, content_type, dumps, loads, default_encoding):
+    def __init__(self, content_type: str, dumps: type_info.DumpSFunction,
+                 loads: type_info.LoadSFunction,
+                 default_encoding: str) -> None:
         self._dumps = dumps
         self._loads = loads
         self.content_type = content_type
         self.default_encoding = default_encoding
 
-    def to_bytes(self, inst_data, encoding=None):
+    def to_bytes(
+            self,
+            inst_data: type_info.Serializable,
+            encoding: typing.Optional[str] = None) -> typing.Tuple[str, bytes]:
         """
         Transform an object into :class:`bytes`.
 
-        :param object inst_data: object to encode
-        :param str encoding: character set used to encode the bytes
+        :param inst_data: object to encode
+        :param encoding: character set used to encode the bytes
             returned from the ``dumps`` function.  This defaults to
             :attr:`default_encoding`
         :returns: :class:`tuple` of the selected content
@@ -100,17 +115,17 @@ class TextContentHandler:
         dumped = self._dumps(escape.recursive_unicode(inst_data))
         return content_type, dumped.encode(selected)
 
-    def from_bytes(self, data, encoding=None):
+    def from_bytes(
+            self,
+            data: bytes,
+            encoding: typing.Optional[str] = None) -> type_info.Deserialized:
         """
         Get an object from :class:`bytes`
 
-        :param bytes data: stream of bytes to decode
-        :param str encoding: character set used to decode the incoming
+        :param data: stream of bytes to decode
+        :param encoding: character set used to decode the incoming
             bytes before calling the ``loads`` function.  This defaults
             to :attr:`default_encoding`
-        :param dict content_parameters: optional :class:`dict` of
-            content type parameters from the :mailheader:`Content-Type`
-            header
         :returns: decoded :class:`object` instance
 
         """